requirements.txt
streamlit
yfinance
plotly
prophet
pandas
ta
newspaper3k
beautifulsoup4
openai
kaleido
gradio
google-generativeai
requests
python-dotenv